Develop

+++ server/src/services/JwtService.js
@@ -5,8 +5,8 @@ require('dotenv').config()
 
 class JwtService {
     generateTokens(payload) {
-        const accessToken = sign(payload, process.env.JWT_ACCESS_SECRET, {expiresIn: '15m'})
-        const refreshToken = sign(payload, process.env.JWT_REFRESH_SECRET, {expiresIn: '30d'})
+        const accessToken = sign(payload, process.env.JWT_ACCESS_SECRET, {expiresIn: process.env.JWT_ACCESS_EXPIRES})
+        const refreshToken = sign(payload, process.env.JWT_REFRESH_SECRET, {expiresIn: process.env.JWT_REFRESH_EXPIRES})
 
         return {
             accessToken,
@@ -16,6 +16,9 @@ class JwtService {
 
     async updateToken(userId, refreshToken) {
         const userToken = await TokensModel.findOne({where: {user_id: userId}})
+        if (!userToken) {
+            return await this.saveNewToken(userId, refreshToken)
+        }
         userToken.token = refreshToken
         await userToken.save()
     }
@@ -30,7 +33,7 @@ class JwtService {
 
     async deleteToken(token) {
         const tokenDB = await TokensModel.findOne({where: {token}})
-        await tokenDB.destroy()
+        tokenDB && await tokenDB.destroy()
     }
 
     validateAccessToken(accessToken) {
